Refactor 6 - Use defs/__init__.py
Summary:
Part 6 of https://github.com/facebookresearch/spdl/pull/1026

Update `defs/__init__.py` and import from there.

Differential Revision: D84880869
