+++ src/spdl/pipeline/_components/_pipe.py
@@ -22,7 +22,7 @@ from typing import Generic, TypeVar
 
 from .._convert import convert_to_async
 from .._utils import create_task
-from ..defs._defs import _PipeArgs
+from ..defs import _PipeArgs
 from ._common import _EOF
 from ._hook import _stage_hooks, _task_hooks, TaskHook
 from ._queue import _queue_stage_hook, AsyncQueue
