+++ src/spdl/pipeline/defs/__init__.py
@@ -26,6 +26,12 @@ The following step explains the steps to build a pipeline using the building blo
 """
 
 from ._defs import (
+    _ConfigBase,
+    _PipeArgs,
+    _PipeType,
+    _TAsyncCallables,
+    _TCallables,
+    _TPipeInputs,
     Aggregate,
     AggregateConfig,
     Disaggregate,
@@ -40,6 +46,12 @@ from ._defs import (
 )
 
 __all__ = [
+    "_ConfigBase",
+    "_PipeArgs",
+    "_PipeType",
+    "_TAsyncCallables",
+    "_TCallables",
+    "_TPipeInputs",
     "Aggregate",
     "AggregateConfig",
     "Disaggregate",
